Invert twin decoding order in DynamicClient
This fixes an issue where v131 twins are decoded as earlier versions.
This is not a permanent fix, and this issue might occur again in the
future. The main problem is that we just try to decode types off all
previously known versions, which is wrong. As seen here, earlier
versions might have an encoding which just so happens to be compatible
and decodable by the new layout, in which case there will obviously be
an error.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>